// src/app/open-api/model/errorData.ts
/**
 * PizzaShop
 * <h1>REST API for managing a pizzeria.</h1> <br>  The application includes the management of:  <ul>     <li>users (workers and customers)</li>     <li>company branches</li>     <li>products (variations like doughs, dimensions and toppings, categories)</li>     <li>orders (creation, management)</li> <ul>
 *
 * OpenAPI spec version: 1.0.0
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ConstraintError } from './constraintError';
import { ErrorInternal } from './errorInternal';

export interface ErrorData { 
    userTitle: string;
    userMessage: string;
    internal?: ErrorInternal;
    constraintErrors?: Array<ConstraintError>;
}